Minor fixes and project restructure:   - fix issue with quantity when add to cart (revert from totalCost as a state, and use prior implementation which is pushing to a list all products with price * quantity)   - fix issue with Fisher-Yates shuffle algorithm. Implement a React.memo() HOC over HomePage to prevent reshuffling in case the user clicks on homePAge link multiple times consecutively   - Restructure project such as all components are now in separate files in a 'component' directory
